Get rid of request validator - folded into request creator
