+++ lib/controllers/MembershipRequestController.js
@@ -1,13 +1,11 @@
 const User = require('../models/User')
 const Band = require('../models/Band')
 const MembershipRequest = require('../models/MembershipRequest');
-const MembershipRequestValidator = require('../helpers/MembershipRequest/MembershipRequestValidator.js')
 const MembershipRequestCreator = require('../helpers/MembershipRequest/MembershipRequestCreator.js')
 const createError = require('http-errors');
 
 const createMembershipRequestAction = async (user, band, res, next) => {
   try {
-    MembershipRequestValidator(user, band)
     await MembershipRequestCreator(user, band)
     res.status(201).send()
   } catch (e) {
